GuidedTours: more reliable element highlighting with scrolling
This commit changes the way we highlight DOM elements for a tour step.
It adds an `Overlay` component that provides a more robust overlay and
is used by the `BasicStep` and `ActionStep` components.

Previously, we created the highlight by drawing a very large box-shadow
around the target element to be highlighted. This worked great at
first, but had a few problems--notably, the highlight wasn't able to
break out of certain containers properly (so not the whole Calypso UI
would be faded out, just e.g. the sidebar); also animation of box-
shadow doesn't seem to be GPU-accelerated. We now draw the highlight
ourselves as a set of four divs (top of the target, left of it, right
of it, and below it) with a certain opacity. That also means that
animating it will likely be much easier on performance.

Previously, the calculation of a step's position, the highlight's
target's position, and the necessary scrolling amount was distributed
across a few places. This resulted in a sort of race condition where
we'd scroll by one amount, but tell the highlight to assume a different
amount (since the scrolling had already started). We now calculate the
step's position and the amount we will need to scroll once in
`componentWillMount` and store the values in the component's state.